Add highlighting for `if` statements:
This only does the line containing the `if` statement itself, the body is not
represented in the tree. This is because parsing the indentation is hard and/or
impossible with tree sitter. Taking a look at the Python tree sitter grammar it
appears that they have a custom parser to handle the indentation, which is too
much work for me at the moment. The important part is that the code looks
pretty, even if the tree sitter AST doesn't 100% reflect the Cicada AST.